feat: Creating custom hook and using it
